# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OpsPilot API
  description: |
    OpsPilot - AI-Powered Infrastructure Operations Platform API.
    All protected endpoints require a Bearer JWT token in the Authorization header.
  version: 2.0.0
  contact:
    name: OpsPilot Team
servers:
  - url: http://localhost:3001
    description: Local development server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        role:
          type: string
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string

    # ── Hosts ─────────────────────────────────────────────────────────
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration successful
        '400':
          description: Invalid input

  # ── Hosts ─────────────────────────────────────────────────────────
...